+++ app/src/lib/components/layout/Header.svelte
@@ -1,24 +1,19 @@
 <script lang="ts">
 	import ButtonWrapper from './ButtonWrapper.svelte';
-	import { API_BASE } from '$lib/global/config';
-	import { page } from '$app/state';
-	import type { SearchResult } from '../../../../../shared/types/core';
-	import SearchResults from '../search/SearchResults.svelte';
 	import Icon from './Icon.svelte';
 	import Logo from './Logo.svelte';
 	import { onMount } from 'svelte';
 
 	let {
 		userIsLoggedIn,
-		menuIsOpen = $bindable()
+		menuIsOpen = $bindable(),
+		searchIsOpen = $bindable()
 	}: {
 		userIsLoggedIn: boolean;
 		menuIsOpen: boolean;
+		searchIsOpen: boolean;
 	} = $props();
 
-	let searchIsOpen = $state(false);
-	let searchQuery = $state('');
-	let searchResults: SearchResult[] = $state([]);
 	let showPWAInstallPrompt = $state(false);
 	let installPromptEvent: any | null = $state(null);
 
@@ -39,32 +34,6 @@
 			showPWAInstallPrompt = true;
 		});
 	});
-
-	// Reset search when page changes
-	$effect(() => {
-		if (page.url) {
-			searchIsOpen = false;
-			searchQuery = '';
-			searchResults = [];
-		}
-	});
-
-	const performSearch = async (query: string) => {
-		if (query.length < 3) {
-			return;
-		}
-		try {
-			const response = await fetch(`${API_BASE}/search?q=${encodeURIComponent(query)}`);
-			if (response.ok) {
-				const data = await response.json();
-				searchResults = data.results;
-			} else {
-				console.error('Search request failed');
-			}
-		} catch (error) {
-			console.error('Error performing search:', error);
-		}
-	};
 </script>
 
 {#if !userIsLoggedIn}
@@ -110,18 +79,6 @@
 	</ButtonWrapper>
 </header>
 
-{#if searchIsOpen}
-	<div class="search-container">
-		<input
-			type="text"
-			placeholder="Search for artists, releases, tracks..."
-			bind:value={searchQuery}
-			onkeydown={(e) => e.key === 'Enter' && performSearch(searchQuery)}
-		/>
-		<SearchResults {searchResults} query={searchQuery} />
-	</div>
-{/if}
-
 <style>
 	.closed-beta-message,
 	.install-prompt {
@@ -153,17 +110,5 @@
 			text-decoration: none;
 			color: inherit;
 		}
-	}
-	input {
-		width: 90%;
-		max-width: 400px;
-		padding: 0.5rem;
-		margin: 0 1rem;
-		border: 1px solid var(--color-accent);
-		border-radius: 4px;
-		font-size: 1rem;
-	}
-	.search-container {
-		width: 100%;
 	}
 </style>
